% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.setup.R
\name{annotate.growth}
\alias{annotate.growth}
\title{Add the growth information onto the known tree.
This uses pplacer to ensure that previously defined clusters remain the same.}
\usage{
annotate.growth(t, t.grown, mc.cores = 1)
}
\arguments{
\item{t:}{An inputted tree using ape's tree handling}

\item{t.growth:}{A set of trees from pplacer. Each with a newly added tip.}

\item{mc.cores:}{A parallel option}
}
\value{
: The input tree annotated with growth information stored as growth.info.
}
\description{
Add the growth information onto the known tree.
This uses pplacer to ensure that previously defined clusters remain the same.
}
