% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.setup.R
\name{annotate.new}
\alias{annotate.new}
\title{Define new sequences}
\usage{
annotate.new(seq.info, which.new)
}
\arguments{
\item{seq.info:}{A data frame or data.table object containing the sequence meta data}

\item{which.new:}{A set of indices of which sequences were to be labelled "new"}
}
\value{
seq.info annotated with a true/false "New" column
}
\description{
Annotate a subset of sequences as "New". This creates an additional column in seq.info
}
