% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.setup.R
\name{annotate.nodes}
\alias{annotate.nodes}
\title{Prepare nodes in tree for clustering}
\usage{
annotate.nodes(t, mc.cores = 1)
}
\arguments{
\item{t:}{An inputted tree passed by extend.tree. Should contain seq.info}

\item{mc.cores:}{A parallel option to increas speed}
}
\value{
A  data table, "node.info" to a annotate the information under a given
node at a tree.
}
\description{
Groups meta data by descendant. This also includes bootstrap certainty referenced
by node labels in the ape object. Unlabeled nodes are defaulted to a bootstrap
value of 1 (maximum certainty), including the root.
}
