% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cluster.ex}
\alias{cluster.ex}
\title{An example set of clusters, built using component.cluster}
\format{
A data.table object with 9 variables:
\describe{
\item{ClusterID}{ The unique identifier number for this cluster. A numberic}
\item{ID}{A list of vectors, each containing the accession IDs (characters) of sequences within a cluster}
\item{CollectionDate}{A list of vectors, each containing the collection date of sequences within a cluster}
\item{Subtype}{A list of vectors, each containing the subtypes (factors) within a cluster}
\item{Header}{A list of vectors, each containing the original headers from the alignement used to build this set of clusters}
\item{Size}{The original size of this cluster before being updated with new cases. This simply the number of sequences within the cluster}
\item{Growth}{The growth of the cluster after new cases are added}
\item{DistThresh}{The pairwise distance threshold used to create this complete set of clusters. Corresponds to a setID as an input parameter}
\item{SetID}{The unique identifier for this set of clusters. A numeric}
}
}
\usage{
cluster.ex
}
\description{
A dataset describing 5 different clusters. Their member headers are listed, as well as the growth they experienced
(ie. the number of new sequences forming clusters with old sequences.). See component.cluster for further information on
how these were assigned based on graph.ex as an input
}
\keyword{datasets}
