% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.clustering.R
\name{component.cluster}
\alias{component.cluster}
\title{Create clusters based on the components of a graph}
\usage{
component.cluster(g, dist.thresh = 0, setID = 0)
}
\arguments{
\item{g:}{The input graph, annotated with vertex, edge, and growth resolution
information}

\item{dist.thresh:}{The maximum distance defining which edges are filtered.
A higher distance threshold implies a larger average cluster size}

\item{setID:}{A numeric identifier for this cluster set.}
}
\value{
A set of clusters as a data.table. See example cluster.ex object
documentation for an example of clustered sequence data + meta data
}
\description{
Edges are filtered away using a distance threshold to break up the completely
connected graph such that only similar edges remain.
}
\examples{
component.cluster.set <- component.cluster(graph.ex, dist.thresh = 0.06)

component.cluster.set[which.max(Size),]
}
