% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-setup.R
\name{extend.tree}
\alias{extend.tree}
\title{Place new sequences on the tree.
This function makes calls to pplacer and guppy binaries (Matsen, 2010)
provided with this package.  pplacer ensures that previously defined clusters
remain the same. New tips are assigned a set of possible placements in the
tree.}
\usage{
extend.tree(phy, seqs.full, log.file = NA, locus = "LOCUS")
}
\arguments{
\item{phy:}{ape::phylo object.  Must be annotated with seq.info, node.info,
and path.info by calling import.tree().}

\item{seqs.full:}{ape::DNAbin or AAbin object.  This sequence alignment
must contain all known sequences in the tree \code{phy}, as
well as all new sequences we want to graft to the tree
as potential cluster growth.}

\item{log.file:}{character.  A path to the logfile from a tree reconstruction
run.  This file can be produced by IQTREE, FastTree or RAxML.}
}
\value{
ape::phylo object with growth.info field
}
\description{
Place new sequences on the tree.
This function makes calls to pplacer and guppy binaries (Matsen, 2010)
provided with this package.  pplacer ensures that previously defined clusters
remain the same. New tips are assigned a set of possible placements in the
tree.
}
