% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{extended.tree.ex}
\alias{extended.tree.ex}
\title{An extension of an ape tree object which can be used to create clusters}
\format{
A phylogenetic tree with 6 tips and 4 internal nodes. Node labels represent
certainty. See ape's implementation of phylogenetic tree objects for information
about tags within this object. In addition, there are 4 new objects created by
functions within tree.setup.R
\describe{
\item{seq.info}{ See seq.info.ex, a data.table containing sequence meta data}
\item{node.info}{ Grouping of the meta.data present in seq.info assigned to
various nodes in the tree, coupled with information important to clustering,
such as mean divergence from root, or node certainty }
\item{path.info}{ Information regarding the path of edges from tips to the root
of the tree. This is also necessary for some clustering algorithms, specifically
step.cluster}
\item{growth.info}{ a data.table pairing new sequences, to a single node in the
tree based on placements assigned by guppy and pplacer. The certainty of this placement,
terminal branch length, neighbour, and branch length from new internal node to
new neighbour are described}
}
}
\usage{
extended.tree.ex
}
\description{
An extension of old.tree.ex maximum likelihood tree built using IQ-TREE with automatic
model selection and 1000 parametric bootstraps. growth information and additional
information useful for cluster identification were added by extend.tree.
}
\keyword{datasets}
