% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-clustering.R
\name{fit.decay}
\alias{fit.decay}
\title{Fit binomial regression model to distribution of bipartite edges between
samples at different time points, as a model of decay in edge density with
time.  (internal)}
\usage{
fit.decay(nodes, edges)
}
\arguments{
\item{nodes:}{data.table, excluding new cases}

\item{edges:}{data.frame, edges below a given distance threshold}
}
\value{
glm object
}
\description{
Fit binomial regression model to distribution of bipartite edges between
samples at different time points, as a model of decay in edge density with
time.  (internal)
}
