% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.data.R
\name{generate.all}
\alias{generate.all}
\title{Generate data found in /data folder}
\usage{
generate.all()
}
\description{
This is partially intended as example use code, however may also act as informal
testing in the development cycle and as a tool to update data quickly if required.
}
