% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get.AIC}
\alias{get.AIC}
\title{Get AIC values from an analysis}
\usage{
get.AIC(cluster.analysis, param.list)
}
\arguments{
\item{cluster.analysis:}{data.table, from some predictive growth model analysis
generated by fit.analysis()}

\item{param.list:}{list, clustering thresholds used to generate cluster
data with multi.cluster()}
}
\value{
The AIC data for all columns containing fit objects. The column specifying
setID is retained
}
\description{
Takes a cluster.analysis and extracts AIC values from columns containing model
fits. Fit columns are automatically identified
}
