% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.setup.R
\name{get.old.seqs}
\alias{get.old.seqs}
\title{Get a subset of sequences based on the old sequences in seq.info}
\usage{
get.old.seqs(seqs, seq.info)
}
\arguments{
\item{seqs:}{A full alaignment}

\item{seq.info:}{A set of seq.info. If provided, may be used to ascertain old sequences from new
A tree is generally only to be built from old sequences}
}
\value{
: Filtered sequences
}
\description{
Get a subset of sequences based on the old sequences in seq.info
}
