% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{graph.ex}
\alias{graph.ex}
\title{An example graph, built based on pairwise TN93 distances}
\format{
A list of 3 variables
\describe{
\item{seq.info}{ See seq.info.ex, a data.table containing sequence meta data}
\item{edge.info}{ A named, weighted pairwise distance matrix. }
\item{growth.resolved}{ a data.table pairing new sequences, to a single sequence.
In order to ensure that clusters do not merge upon growth, a new sequence may only
join up to one old cluster. By default, new sequences join the cluster of the
sequence they are most similar to. }
}
}
\usage{
graph.ex
}
\description{
This implementation of a graph is a list, describing a set of sequences and the
distances between them. See create.graph for more information on how this graph
was created using alignment.ex as input
}
\keyword{datasets}
