% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-setup.R
\name{import.tree}
\alias{import.tree}
\title{Prepare a tree for clustering analysis}
\usage{
import.tree(phy, seq.info = data.table())
}
\arguments{
\item{phy:}{ape::phylo object.  The tree should only contain "known"
sequences, i.e., none of the "New" sequences we want to predict.}

\item{seq.info:}{data.table object.  Contains metadata for both known and
new sequences, where the latter are identified by not
appearing in the input tree \code{phy}.  Must contain sequence
labels under column label \code{Header}.}
}
\value{
The tree annotated with node information and seq.info
}
\description{
Prepare a tree for clustering analysis
}
