% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-clustering.R
\name{mono.pat.cluster}
\alias{mono.pat.cluster}
\title{Obtain monophyletic clusters based on pairwise distances}
\usage{
mono.pat.cluster(
  t,
  dist.thresh,
  boot.thresh = 0,
  dist.criterion = "max.patristic.dist",
  setID = 0
)
}
\arguments{
\item{t:}{The input tree file, annotated with vertex and edge information}

\item{dist.criterion:}{A particular column in node.info that must be less than
a distance threshold. By default, this is "max.patristic.dist", however, the
"mean.patristic.dist" is also a column. Other columns added to node.info can
be checked against dist.thresh, however, these would be added by the user.}

\item{dist.thresh:}{The threshold required for clustering. Monophyletic groups
with criterion under this value could be considered clusters.}

\item{setID:}{A numeric identifier for this cluster set.}

\item{boot.thresh:}{The minimum bootstrap threshold defining clusters. Monophyletic
groups with a parent node more certain than this criterion could be considered
clusters}
}
\value{
: A set of clusters as a data.table. See example cluster.ex object
documentation for an example of clustered sequence data + meta data
}
\description{
Clusters are defined as  as a monophyletic clade under a high-confidence common
ancestor. Some measure of divergence (criterion) within this clade must fall
under a distance threshold in order for it to be labelled a cluster.
}
