% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{multi.cluster}
\alias{multi.cluster}
\title{Wrapper for cluster.method functions}
\usage{
multi.cluster(obj, param.list, cluster.method, rangeID = 0)
}
\arguments{
\item{obj:}{S3 object of class clusData}

\item{param.list:}{list, each entry should be a list of threshold values
to pass to the clustering method}

\item{cluster.method:}{function that returns data frame of known cases
annotated with cluster ID and growth}

\item{rangeID:}{integer, unique index for rows generated by this function}
}
\value{
list
}
\description{
Wrapper for cluster.method functions
}
