% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence-setup.R
\name{parse.headers}
\alias{parse.headers}
\title{Convenience function to parse metadata from sequence headers}
\usage{
parse.headers(
  headers,
  var.names = c(),
  var.transformations = list(),
  sep = "_"
)
}
\arguments{
\item{headers:}{character, vector of sequence names such as obtained by
names(x) where x is an ape::DNAbin object.}

\item{var.names:}{character, The names of the variables represented in each header.}

\item{var.transformations:}{list, A list of transformation functions (such as as.Date())
these transform each row into it's proper type.  By default, each value is taken
literally (i.e., as.character).}

\item{sep:}{character, The separator character that splits up the headers into
values.}
}
\value{
A data.table object containing sequence meta data paired to headers.
See seq.info.ex, for more details and an example
}
\description{
Translates a set of sequence headers into a data.frame object for data input.
Requires that every sequence header contains the same number and type of
values, separated by a reserved delimiter character, e.g., underscore "_"
}
\examples{
seq.info <- parse.headers(names(alignment.ex), var.names = c("ID", "CollectionDate", "Subtype"),
                         var.transformations = list(as.character, as.Date, as.factor))
}
