% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot.aic.diff}
\alias{plot.aic.diff}
\title{Plots the difference in AIC across the proposed and null models
This will reach a central optima, as extremes produce low AIC differences
Greater negative values mean larger improvement relative to null model}
\usage{
\method{plot}{aic.diff}(res)
}
\arguments{
\item{res:}{The result of a fit.analysis() run.}
}
\value{
: A set of AIC differences.
}
\description{
Plots the difference in AIC across the proposed and null models
This will reach a central optima, as extremes produce low AIC differences
Greater negative values mean larger improvement relative to null model
}
