% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.setup.R
\name{pull.headers}
\alias{pull.headers}
\title{Get seq.info from headers}
\usage{
pull.headers(seqs, var.names, var.transformations = list(), sep = "_")
}
\arguments{
\item{seqs:}{An inputted alignment using ape's sequence handling}

\item{var.names:}{The names of the variables represented in each header.}

\item{var.transformations:}{A list of transformation functions (such as as.Date())
these transform each row into it's proper type. by default, each type is character.}

\item{sep:}{The separator character that splits up the headers into variables}
}
\value{
A data.table object containing sequence meta data paired to headers.
See seq.info.ex, for more details and an example
}
\description{
Translates a set of sequence headers into a data.frame object for data input.
For well formatted alignments, this may be an easy way to obtain meta-data.
}
\examples{
seq.info <- pull.headers(alignment.ex, var.names = c("ID", "CollectionDate", "Subtype"),
                         var.transformations =list(as.character, as.Date, as.factor))
}
