% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.setup.R
\name{run.pplacer_guppy}
\alias{run.pplacer_guppy}
\title{Get grown phylogenies}
\usage{
run.pplacer_guppy(refpkg)
}
\arguments{
\item{refpkg:}{A reference package to use as input for pplacer}
}
\value{
A set of trees, each containing 1 new sequence.
}
\description{
A wrapper for pplacer's basic run function coupled with guppy's sing function.
Together, these extend fixed trees with most likely placement locations.
}
