% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.setup.R
\name{run.tn93}
\alias{run.tn93}
\title{A wrapper for tn93's basic run function to get an edgelist
NOTE: The sequences referenced here will also be referenced in another data set (seq.info)}
\usage{
run.tn93(seqs.full)
}
\arguments{
\item{seqs.full:}{The full alignment. Including sequences excluded from the tree.}
}
\value{
: An edgelist of pairwise TN93 distances calculated using tn93 binaries
}
\description{
A wrapper for tn93's basic run function to get an edgelist
NOTE: The sequences referenced here will also be referenced in another data set (seq.info)
}
