% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seq.info.ex}
\alias{seq.info.ex}
\title{An example set of sequence meta.data corresponding to alignment.ex}
\format{
A data.table object with 9 variables:
\describe{
\item{ID}{Accession IDs (characters) of sequences}
\item{CollectionDate}{Collection date of sequences. Full dates given as yyyy-mm-dd}
\item{Subtype}{Subtypes (factors) within a cluster}
\item{Header}{The original headers from the alignement. This matches meta data to sequences}
}
}
\usage{
seq.info.ex
}
\description{
A dataset describing 10 different HIV1 pol sequences collected in Northern Alberta Canada.
}
\keyword{datasets}
