% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pplacer.utils.R
\name{translate.log}
\alias{translate.log}
\title{Parses the logfile output of tree building software to find information relevant to pplacer.
Prints stats to a temporary ref.pkg file and returns the path to that file.}
\usage{
translate.log(log.file, program, substitution.model = "GTR")
}
\arguments{
\item{log.file:}{A path to the logfile from a tree construction run}

\item{program:}{The software used to build the tree.
Can be "FastTree", "RAxML" or "IQ-TREE".}

\item{substitution.model:}{The substitution model used. Currently only accepts "GTR"}
}
\value{
: a json output to be written to a given stats.file for pplacer
}
\description{
NOTE: Currently compatible with GTR substitution model and either FastTree or IQ-TREE logfiles.
Working to extend this to PhyML logfiles, and then to different models of evolution
}
