% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-setup.R
\name{annotate.growth}
\alias{annotate.growth}
\title{Add the growth information onto the known tree.}
\usage{
annotate.growth(phy, ptree)
}
\arguments{
\item{phy:}{ape::phylo object.  Return value of extend.tree().}

\item{ptree:}{ape::phylo object.  A tree generated by pplacer with a
newly added tip placed by maximum likelihood.}
}
\value{
data.table. The growth.info for a given tree. This includes the
branch length to the newly added terminal node, whether or not the
neighbour to the newly added terminal node is also terminal, and the
branch length to that neighbour, from the newly placed internal node
(the "PendantLength")
}
\description{
Add the growth information onto the known tree.
}
