% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.setup.R
\name{annotate.nodes}
\alias{annotate.nodes}
\title{Called by extend.tree. Adds additional node info to the tree. Required for mono.cluster()
NOTE: Unlabeled nodes are defaulted to a bootstrap value of 1}
\usage{
annotate.nodes(t, mc.cores = 1)
}
\arguments{
\item{t:}{An inputted tree using ape's tree handling. This must be annotated with seq.info}

\item{mc.cores:}{A parallel option}
}
\value{
: A slightly more detailed data table to replace "node.info" to a given tree.
}
\description{
Called by extend.tree. Adds additional node info to the tree. Required for mono.cluster()
NOTE: Unlabeled nodes are defaulted to a bootstrap value of 1
}
