% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-setup.R
\name{annotate.paths}
\alias{annotate.paths}
\title{Get paths relative to starting nodes}
\usage{
annotate.paths(phy)
}
\arguments{
\item{phy:}{ape::phylo object. An inputted tree An inputted tree passed by
extend.tree.}
}
\value{
list of N matrices, where N is the number of nodes in the input tree.
Each matrix has dimension 3 x P, where P is the path length from
the node to the root (inclusive).  The three rows are:
- Node: integer index of node in path.
- Boot: bootstrap support associated with the node
- BranchLength: length of branch associated with the node
For the root, no branch length is given (NA) and Boot is set to 1.
}
\description{
Called by extend.tree. Adds path info to the tree, ie. The path of branches from
each node to the root. Required for step.cluster()
}
