% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-clustering.R
\name{component.cluster}
\alias{component.cluster}
\title{Create clusters based on the components of a graph}
\usage{
component.cluster(obj, dist.thresh = 0, setID = 0, time.var = NA)
}
\arguments{
\item{obj:}{S3 object of class clusData.  Contains vertex, edge and growth
resolution information.}

\item{dist.thresh:}{double, the maximum distance defining which edges are
filtered. A higher distance threshold implies a larger average cluster size}

\item{setID:}{A numeric identifier for this cluster set.}

\item{time.var:}{character, column name for discrete time variable to fit
a model of edge density decay with time (optional)}
}
\value{
data.frame, known cases annotated with cluster ID and growth
}
\description{
Edges are filtered away using a distance threshold to break up the completely
connected graph such that only similar edges remain.
}
\examples{
component.cluster.set <- component.cluster(graph.ex, dist.thresh = 0.06)

component.cluster.set[which.max(Size),]
}
