% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.clustering.R
\name{component.cluster}
\alias{component.cluster}
\title{Create clusters based on the components of a graph}
\usage{
component.cluster(g, dist.thresh = 0.007, setID = 0)
}
\arguments{
\item{g:}{The input graph, annotated with vertex and edge information}

\item{dist.thresh:}{The maximum distance defining which edges are filtered}

\item{setID:}{If several different parameter ranges are used, the setID can identify them}
}
\value{
A data table which represents cluster information. This includes growth info
Because data.tables are being used, this prevents original values being reassigned via pointer
}
\description{
This uses a homogenization algorithm to identify disconnected components in a graph.
Edges are filtered away using a distance threshold to result in components.
}
\examples{
load("data/graph_ex.RData")
component.cluster.set <- component.cluster(g, dist.thresh = 0.015)
}
