% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.setup.R
\name{create.graph}
\alias{create.graph}
\title{Makes a graph object based on sequence data}
\usage{
create.graph(
  seq.info,
  edge.info,
  growth.resolution = minimum.retrospective.edge
)
}
\arguments{
\item{seq.info:}{A set of sequence meta-data sorted by alignment header}

\item{edge.info:}{A pairwise edge matrix of all associated headers in seq.info}

\item{growth.resolution:}{The method by which growth is resolved. This ensures new cases don't merge clusters
By default, each new sequence joins a cluster byonly it's minimum retrospective}
}
\value{
A graph, with sequences and edge info. New sequences are only linked by their minimum retrospective edge
}
\description{
Create an implementation of a graph. This consists of an edge matrix (see ape::dist.dna()) and some sequence meta data
A large part of this process involves resolving growth, ensuring that new sequences are only added prospectively without merging clusters.
At this point, we also annotate the minimum retrospective edges of each edge. This is stored in sequence information
}
