% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.setup.R
\name{extend.tree}
\alias{extend.tree}
\title{Prepare a tree for clustering}
\usage{
extend.tree(
  t,
  seq.info,
  mc.cores = 1,
  log.file = NA,
  full.align = character(0),
  locus = "LOCUS"
)
}
\arguments{
\item{t:}{An inputted tree using ape's tree handling}

\item{seq.info:}{A data frame or data.table object containing the sequences.
By default, new sequences are assigned as those not in the tree.}

\item{log.file:}{A path to the logfile from a tree construction run}

\item{seqs.full:}{The full alignment. Including sequences excluded from the tree.}

\item{mc.cores:}{Passed to annotate.nodes as a parallel option}
}
\value{
The tree annotated with node information and seq.info
}
\description{
Extends an ape tree object to include annotated node and path information.
By default this will include some additional info like paths, bootstraps and growth.
}
\examples{
t <- extend.tree(old.tree.ex, seq.info.ex,
                 full.align = alignment.ex,
                 log.file = "data/IQTREE_log_ex.txt")

step.cluster(t)
}
