% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.setup.R
\name{extend.tree}
\alias{extend.tree}
\title{Extends an ape tree object to include annotated node and path information.
By default this will include some basic node info like bootstraps. Other annotations are optional.}
\usage{
extend.tree(t, seq.info, paths.annotated = T, mc.cores = 1)
}
\arguments{
\item{t:}{An inputted tree using ape's tree handling}

\item{seq.info:}{A data frame or data.table object containing the sequences.
By default, new sequences are assigned as those not in the tree.}

\item{mc.cores:}{Passed to annotate.nodes as a parallel option}

\item{paths.annotated:}{An option to add path info for the tree. Useful for some tree-based clustering methods}
}
\value{
: the tree annotated with node information and seq.info
}
\description{
Extends an ape tree object to include annotated node and path information.
By default this will include some basic node info like bootstraps. Other annotations are optional.
}
