% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{fit.analysis}
\alias{fit.analysis}
\title{Predictive analysis on clusters}
\usage{
fit.analysis(
  cluster.data,
  mc.cores = 1,
  predictor.transformations = list(),
  predictive.models = list(NullModel = function(x) {     glm(Size ~ Growth, data = x,
    family = "poisson") })
)
}
\arguments{
\item{cluster.data:}{Inputted set(s) of clusters. Possibly multiple ranges}

\item{mc.cores:}{A parallel option to increase run speed}

\item{predictor.transformations:}{A named list of transformation functions for
each predictor variable (ex. list("Data"==sum). Because clustered meta data takes
the form of a list these functions are often necessary to obtain a single,
cluster-level variable}

\item{predictive.models:}{A named list of functions, each of which applies a
model to inputted cluster data (x). By default a "NullModel" example. Where
Growth is predicted only by cluster size}
}
\value{
A data.table of analysis results. Model fits are stored as entries in
the rows of a data.table. The column specifying setID is retained, as is the
range ID and the parameters used to create the cluster.
}
\description{
Fits predictive model of some outcome (by default, cluster growth) to some
cluster-level variable (by default, cluster size). This fit is done for each
cluster set. Multiple models can be inputted as a named list of functions taking
in cluster data (see example)
}
\examples{
cluster.data <- cluster.ex
cluster.data[,"RangeID":=0]

fit.result <- fit.analysis(cluster.data)

mod.performance <- fit.result$NullModel
}
