% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get.AIC}
\alias{get.AIC}
\title{Get AIC differences from an analysis}
\usage{
get.AIC(cluster.analysis)
}
\arguments{
\item{cluster.analysis:}{A data.table from some predictive growth model analysis}
}
\value{
The AIC data for all columns containing fit objects
}
\description{
Takes a cluster.analysis and extracts AIC values from columns containing models
}
