% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get.AIC}
\alias{get.AIC}
\title{Get AIC values from an analysis}
\usage{
get.AIC(cluster.analysis)
}
\arguments{
\item{cluster.analysis:}{A data.table from some predictive growth model analysis
generated by fit.analysis()}
}
\value{
The AIC data for all columns containing fit objects. The column specifying
setID is retained
}
\description{
Takes a cluster.analysis and extracts AIC values from columns containing model
fits. Fit columns are automatically identified
}
\examples{
predictive.models <- list(
  "NullModel" = function(x){
    glm(Growth~Size, data=x, family="poisson")
  },
  "TimeModel" = function(x){
    glm(Growth~Size+CollectionDate, data=x, family="poisson")
  }
)

predictor.transformations <- list(
  "CollectionDate" = function(x){mean(x)}
)

cluster.data <- cluster.ex
cluster.data[,"RangeID":=0]

res <- fit.analysis(cluster.data,
                    predictor.transformations = predictor.transformations,
                    predictive.models = predictive.models)

aics <- get.AIC(res)
aicdiff <- aics$TimeModelAIC-aics$NullModelAIC
which.min(aicdiff)
}
