% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.setup.R
\name{get.old.seqs}
\alias{get.old.seqs}
\title{Obtain "old" sequences}
\usage{
get.old.seqs(seqs, seq.info)
}
\arguments{
\item{seqs:}{A full alignment}

\item{seq.info:}{A set of seq.info. If provided, may be used to ascertain old sequences from new
A tree is generally only to be built from old sequences}
}
\value{
Filtered sequences, an ape seq object
}
\description{
Get a subset of sequences from an ape seq object based on seq.info data which is not labeled new
}
