% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-setup.R
\name{minimum.retrospective.edge}
\alias{minimum.retrospective.edge}
\title{The default growth resolution helper.}
\usage{
minimum.retrospective.edge(obj)
}
\arguments{
\item{obj:}{S3 object of class clusData from create.graph}
}
\value{
integer, vector of row indices into g$edge.info
}
\description{
Ensures that new sequences only join old clusters through the shortest
retrospective edge, i.e., an edge connecting the new node to a non-new node.
}
