% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.setup.R
\name{minimum.retrospective.edge}
\alias{minimum.retrospective.edge}
\title{The default growth resolution helper.}
\usage{
minimum.retrospective.edge(g)
}
\arguments{
\item{g:}{The input graph}
}
\value{
A data table matching each new sequence with its closest neighbour
}
\description{
Ensures that new sequences only join old clusters through their minimum
retrospective edge. A growth resolution
}
