% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.clustering.R
\name{mono.pat.cluster}
\alias{mono.pat.cluster}
\title{Clusters as a monophyletic clade under a high-confidence common ancestor.
The pairwise patristic distances in this clade must all
NOTE: Running this method requires a tree object with growth.info defined}
\usage{
mono.pat.cluster(
  t,
  dist.thresh,
  boot.thresh = 0,
  dist.criterion = "max.patristic.dist",
  verbose = T,
  setID = 0
)
}
\arguments{
\item{t:}{The input tree file, annotated with vertex and edge information}

\item{dist.criterion:}{A particular column in node.info that must be less than a distance threshold}

\item{dist.thresh:}{The threshold required for clustering.}

\item{verbose:}{An output monitoring option}

\item{setID:}{If several different parameter ranges are used, the setID can identify them.}

\item{boot.thresh:}{The minimum bootstrap criterion defining clusters}
}
\value{
: A data table which extends a subset of node.info. This includes growth info
}
\description{
Clusters as a monophyletic clade under a high-confidence common ancestor.
The pairwise patristic distances in this clade must all
NOTE: Running this method requires a tree object with growth.info defined
}
