% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{multi.cluster}
\alias{multi.cluster}
\title{Multiple clusters from a parameter set}
\usage{
multi.cluster(
  cluster.method,
  param.list,
  mc.cores = 1,
  verbose = T,
  rangeID = 0
)
}
\arguments{
\item{param.list:}{A named list of parameter sets. Each must correspond to the clustering method used.}

\item{rangeID:}{If several different parameter ranges are used, the rangeID can identify them.}

\item{mc.cores:}{A parallel option}
}
\value{
A larger data.table with parameter sets noted
}
\description{
Runs a given clustering method over a range of parameters values.
If growth is attached This result can be fed into a fit.analysis
}
