% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{multi.cluster}
\alias{multi.cluster}
\title{Multiple clusters from a parameter set}
\usage{
multi.cluster(cluster.method, param.list, mc.cores = 1, rangeID = 0)
}
\arguments{
\item{cluster.method:}{A given clustering function such as step.cluster() which
produces a set of clusters.}

\item{param.list:}{A list, each entry a named list of parameter sets which can
act as inputs to the cluster.method. These include values such as trees and graphs,
as well as  criteria for clustering such as boot.thresh or dist.thresh.}

\item{rangeID:}{A unique identifier for the set of rows generated by this run
if this output is bound to other cluster ranges in a larger analysis.}

\item{mc.cores:}{A parallel option to increase run speed.}
}
\value{
A data.table of clusters. Multiple cluster sets are collected into a range.
}
\description{
Runs a given clustering method (a passed function) over a range of parameter
values (a list, each entry a named list of parameters for the function).
Collects the data into a single data table with multiple cluster set ID's
indicating the parameter set used to define clusters a unique cluster range ID.
}
\examples{
param.list.boot <- lapply(c(0.005,0.01), function(x){
  list("t"=extended.tree.ex, "branch.thresh"=x, "boot.thresh"=0.7)
})

step.cluster.range <- multi.cluster(step.cluster, param.list.boot, mc.cores = 1)
}
