% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{multi.cluster}
\alias{multi.cluster}
\title{Multiple clusters from a parameter set}
\usage{
multi.cluster(cluster.method, param.list, mc.cores = 1, rangeID = 0)
}
\arguments{
\item{cluster.method:}{A given clustering function such as step.cluster() which produces a set of clusters}

\item{param.list:}{A named list of parameter sets, which can act as inputs to cluster.method.}

\item{rangeID:}{A unique identifier for the set of rows generated by this run.
If this output is bound to other cluster ranges in a larger analysis, this can disambiguate}

\item{mc.cores:}{A parallel option to increase run speed.}
}
\value{
A data.table with parameter sets and cluster IDs specified
}
\description{
Runs a given clustering method over a range of parameters values to output
a range of cluster sets corresponding to different
}
\examples{
param.list.boot <- lapply(c(0.005,0.01), function(x){
  list("t"=extended.tree.ex, "branch.thresh"=x, "boot.thresh"=0.7)
})

step.cluster.range <- multi.cluster(step.cluster, param.list.boot, mc.cores = 1)
}
