% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot.aic.diff}
\alias{plot.aic.diff}
\title{Plot the difference in AIC across the proposed and null models}
\usage{
\method{plot}{aic.diff}(res)
}
\arguments{
\item{res:}{The result of a fit.analysis() run.}
}
\value{
A set of AIC differences.
}
\description{
A visualization tool to see the AIC of models with respect to cluster sets
The difference from a null model will reach a central optima, as extremes produce low AIC differences
Greater negative values mean larger improvement relative to a null model
}
