% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.setup.R
\name{pull.headers}
\alias{pull.headers}
\title{Get seq.info from headers}
\usage{
\method{pull}{headers}(seqs, var.names, var.transformations = list(), sep = "_")
}
\arguments{
\item{seqs:}{An inputted alignment using ape's sequence handling}

\item{var.names:}{The names of the variables represented in each header. This must contain "ID".}

\item{var.transformations:}{A list of transformation functions (such as as.character())
these transform each row into it's proper type. by default, each type is set to character.}

\item{sep:}{The separator character that splits upthe headers in the fasta file}
}
\value{
A data.table object containing the information associated with each sequence.
}
\description{
Translates a set of sequence headers into a data.frame object for data input.
NOTE: This must contain, at minimum, a set of unique sequence id's (labelled ID) and one other variable
}
