% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seq.info.ex}
\alias{seq.info.ex}
\title{An example set of sequence meta.data corresponding to alignment.ex}
\format{
A data.table object with 4 variables:
\describe{
\item{ID}{Accession IDs (characters) of sequences}
\item{CollectionDate}{Collection date of sequences. Full dates given as yyyy-mm-dd}
\item{Subtype}{Subtypes (factors) of sequences}
\item{Header}{The original headers from the alignement. This matches meta
data to sequences in original alignment}
}
}
\usage{
seq.info.ex
}
\description{
Built from alignment.ex, the example 10 sequence alignment using parse.headers.
The date of each sequence's collection, it's genbank unique accession ID, and
sequence subtype are referenced within the header
}
\keyword{datasets}
