% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.clustering.R
\name{step.cluster}
\alias{step.cluster}
\title{Obtain clusters based on path to ancestor in tree}
\usage{
step.cluster(t, branch.thresh = 0.03, boot.thresh = 0, setID = 0)
}
\arguments{
\item{t:}{The input tree file, annotated with vertex and edge information}

\item{branch.thresh:}{The maximum branch length criterion defining clusters}

\item{boot.thresh:}{The minimum bootstrap criterion defining clusters}

\item{setID:}{If several different parameter ranges are used, the setID can identify them}
}
\value{
A data table which extends a subset of node.info. This includes growth info
}
\description{
Clusters are defined as a series of tips diverging from A high confidence common ancestor.
This divergence must be done through a series of short branches, which the branch.thresh constrains.
}
\examples{
step.cluster.set <- step.cluster(extended.tree.ex, branch.thresh = 0.03, boot.thresh = 0)

step.cluster.set[which.max(Size), ]
}
