% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-clustering.R
\name{step.cluster}
\alias{step.cluster}
\title{Obtain paraphyletic clusters}
\usage{
step.cluster(obj, branch.thresh = 0.03, boot.thresh = 0, setID = 0)
}
\arguments{
\item{obj:}{S3 object of class "phylo". The input tree file, extended to be annotated with vertex, edge and
growth information.}

\item{branch.thresh:}{The maximum branch length criterion defining clusters.
A branch exceeding this value separates the tip from it's ancestor's cluster.
Higher values imply larger average cluster sizes.}

\item{boot.thresh:}{The minimum bootstrap criterion defining clusters.
Lower values imply larger average cluster size}

\item{setID:}{A numeric identifier for this cluster set.}
}
\value{
: A set of clusters as a data.table. See example cluster.ex object
documentation for an example of clustered sequence data + meta data
}
\description{
Clusters are defined as a group of tips diverging from a high confidence common
ancestor. Divergence must be done through a series of short branches, which
the branch.thresh constrains.
}
\examples{
step.cluster.set <- step.cluster(extended.tree.ex, branch.thresh = 0.03, boot.thresh = 0)

step.cluster.set[which.max(Size), ]
}
