% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-setup.R
\name{translate.log}
\alias{translate.log}
\title{Build a stats.json}
\usage{
translate.log(log.file)
}
\arguments{
\item{log.file:}{A path to the logfile from a tree construction run.}
}
\value{
A json output to be written to a given stats.file for pplacer
}
\description{
Parses the logfile output of tree building software to find information relevant
to pplacer. Prints stats to a temporary ref.pkg file and returns the file path
NOTE: Currently compatible with GTR substitution model and either FastTree,
RAxML or IQ-TREE logfiles. Working to extend this to PhyML logfiles, and to
different models of evolution
}
